requirements.txt
contextlib2==0.5.5
Django==1.11.3
django-cors-middleware==1.3.1
django-environ==0.4.3
django-extensions==1.8.1
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
psycopg2==2.7.3
PyJWT==1.5.2
pytz==2017.2
raven==6.1.0
six==1.10.0
